+++ packages/uniapp-request/src/utils/config.ts
@@ -101,6 +101,11 @@ export const useConfig = (config: Config) => {
          */
         tokenStorageKeyName: config.tokenStorageKeyName ?? '',
         /**
+         * 自定义获取task_id处理程序，通过promise返回最新task_id值即可
+         * + `1.5.11` 及以上版本支持
+         */
+        taskIdValue: config.taskIdValue,
+        /**
          * 自定义获取token处理程序，通过promise返回最新token值即可
          * + `1.0.2` 及以上版本支持
          * @returns 
@@ -149,6 +154,14 @@ export const useConfig = (config: Config) => {
          * 自定义token失效的错误代码，便于请求库内部做自动刷新token判断
          */
         tokenExpiredCode: config.tokenExpiredCode ?? 403,
+        /**
+         * token失效错误代码类型，支持 `httpStatusCode` 和 `apiResponseCode`，默认为 `httpStatusCode`
+         * + `httpStatusCode`: 原生http请求状态码
+         * + `apiResponseCode`: 接口响应错误码
+         * 
+         * + `1.5.11` 及以上版本支持
+         */
+        tokenExpiredCodeType: config.tokenExpiredCodeType ?? 'httpStatusCode',
         /**
          * 请求失败是否自动重试
          */